use barrel import, make CameraSelector a separate widget
