+++ lib/animoji_intro/widgets/animoji_intro_body.dart
@@ -1,8 +1,5 @@
-import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:holobooth/animoji_intro/animoji_intro.dart';
-import 'package:holobooth/camera/bloc/camera_bloc.dart';
 import 'package:holobooth/camera/camera.dart';
 import 'package:holobooth/l10n/l10n.dart';
 import 'package:holobooth_ui/holobooth_ui.dart';
@@ -120,79 +117,10 @@ class _BottomContent extends StatelessWidget {
               textAlign: smallScreen ? TextAlign.center : TextAlign.left,
             ),
           ),
-          _CameraSelector(),
-          if (smallScreen) const SizedBox(height: 16),
+          const CameraSelector(),
           const Flexible(child: AnimojiNextButton()),
         ],
       ),
-    );
-  }
-}
-
-class _CameraSelector extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder(
-      future: availableCameras(),
-      builder: (_, snapshot) {
-        if (snapshot.hasData) {
-          final cameras = snapshot.data;
-
-          if (cameras == null || cameras.isEmpty) {
-            return const Icon(
-              key: Key('animojiIntro_no_camera_icon'),
-              Icons.videocam_off,
-              color: HoloBoothColors.white,
-            );
-          }
-
-          final dropdownItems = cameras
-              .map(
-                (camera) => DropdownMenuItem(
-                  value: camera,
-                  child: Text(
-                    camera.name,
-                    style: Theme.of(context)
-                        .textTheme
-                        .bodyLarge
-                        ?.copyWith(color: HoloBoothColors.white),
-                  ),
-                ),
-              )
-              .toList();
-
-          context.read<CameraBloc>().add(CameraChanged(cameras[0]));
-
-          return BlocBuilder<CameraBloc, CameraState>(
-            builder: (context, state) => DropdownButton<CameraDescription>(
-              borderRadius: const BorderRadius.all(Radius.circular(12)),
-              value: state.camera,
-              dropdownColor: HoloBoothColors.darkPurple,
-              items: dropdownItems,
-              onChanged: (value) =>
-                  context.read<CameraBloc>().add(CameraChanged(value)),
-            ),
-          );
-        }
-
-        if (snapshot.hasError) {
-          final error = snapshot.error;
-          if (error is CameraException) {
-            return CameraErrorView(
-              key: const Key('animojiIntro_camera_error_view'),
-              error: error,
-              textStyle: Theme.of(context)
-                  .textTheme
-                  .bodySmall
-                  ?.copyWith(color: HoloBoothColors.white),
-            );
-          }
-        }
-
-        return const CircularProgressIndicator(
-          color: HoloBoothColors.convertLoading,
-        );
-      },
     );
   }
 }
