+++ lib/camera/camera.dart
@@ -1,1 +1,2 @@
+export 'bloc/camera_bloc.dart';
 export 'widgets/widgets.dart';
