+++ lib/camera/widgets/widgets.dart
@@ -1,2 +1,3 @@
 export 'camera_error_view.dart';
+export 'camera_selector.dart';
 export 'camera_view.dart';
