+++ lib/photo_booth/widgets/photobooth_body.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:holobooth/audio_player/audio_player.dart';
 import 'package:holobooth/avatar_detector/avatar_detector.dart';
-import 'package:holobooth/camera/bloc/camera_bloc.dart';
 import 'package:holobooth/camera/camera.dart';
 import 'package:holobooth/footer/footer.dart';
 import 'package:holobooth/in_experience_selection/in_experience_selection.dart';
